// src/app/angular/types.ts
export interface Character {
  id: number;
  name: string;
}

export interface User {
  name: string;
  email: string;
  username: string;
  password: string;
}

export interface Account {
  id: string;
  name: string;
  email: string;
}

export interface Game {
  id: number;
  name: string;
  description: string;
}